refactor: separate methods for explore interactions
